+++ TimeManager/src/Application/Projects/CancelProject/CancelProjectCommandHandler.cs
@@ -1,11 +1,12 @@
 ﻿using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
+using TimeManager.Application.Common.Models;
 using TimeManager.Domain.Projects;
 
 namespace TimeManager.Application.Projects.CancelProject
 {
-    public class CancelProjectCommandHandler : IRequestHandler<CancelProjectCommand, Unit>
+    public class CancelProjectCommandHandler : IRequestHandler<CancelProjectCommand, Response>
     {
         private readonly IProjectRepository _repository;
 
@@ -14,11 +15,20 @@ namespace TimeManager.Application.Projects.CancelProject
             _repository = repository;
         }
 
-        public async Task<Unit> Handle(CancelProjectCommand request, CancellationToken cancellationToken)
+        public async Task<Response> Handle(CancelProjectCommand request, CancellationToken cancellationToken)
         {
+            var response = new Response();
+
+            var entity = await _repository.GetByIdsAsync(request.ProjectId);
+            if (entity == null)
+            {
+                response.AddError("Projeto não encontrado");
+                return response;
+            }
+
             await _repository.DeleteAsync(request.ProjectId);
 
-            return Unit.Value;
+            return new Response(Unit.Value);
         }
     }
 }
